feat(trade): removed the accounts state since its in services
