--- admin/trade/src/state/accounts/index.ts
@@ -1,8 +1,0 @@
-/** @format */
-
-import accounts from './accounts';
-
-export const { accountAdded, accountDeleted } = accounts.actions;
-export default accounts.reducer;
-export * from './accountsSelectors';
-export * from './accounts';
