+++ admin/trade/src/state/index.ts
@@ -4,7 +4,6 @@ import store from './state';
 export * from './operations';
 export * from './currencyGroups';
 export * from './state';
-export * from './accounts';
 export * from './transactions';
 
 export default store;
